Prepare support for 32-bit mode flags
Since Linux 3.15 (11 years ago) the size of the mode field was
increased from 8 to 32 bits. In order to support some of the newer
mode flags (above the first 8 bits), let's switch to use the new
ioctls and store a 32 bit value in SpiDevObject.

This breaks the compatibility with Linux kernel versions before
v3.15.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>